style: 상단 컴포넌트 분리
- 세부사항 페이지의 상단을 구성하는 그래프와 세부 버튼에 대한 컴포넌트를 페이지에서 분리
